v12

+++ API/Controllers/UserController.cs
@@ -1,7 +1,6 @@
-﻿using CarGuideDDD.Core.MapObjects;
-using CarGuideDDD.Infrastructure.Services;
+﻿using CarGuideDDD.Core.DtObjects;
+using CarGuideDDD.Core.MapObjects;
 using CarGuideDDD.Infrastructure.Services.Interfaces;
-using DTOs;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 
@@ -22,18 +21,13 @@ namespace API.Controllers
         [HttpGet("GetUsers")]
         public async Task<IActionResult> GetUsers()
         {
-
-
-            var users = await _userService.GetAllUsersAsync();
-            return Ok(users);
+            return Ok( await _userService.GetAllUsersAsync());
         }
 
         [Authorize(AuthenticationSchemes = "Bearer", Roles = "Admin")]
         [HttpPost("CreateUser")]
         public async Task<IActionResult> CreateUser([FromBody] UserDto user)
         {
-
-
             return await _userService.AddUserAsync(user);
         }
 
@@ -50,33 +44,25 @@ namespace API.Controllers
         [HttpDelete("DeleteUser")]
         public async Task<IActionResult> DeleteUser([FromBody] UsernameDto user)
         {
-
-
-            return await _userService.DeleteUserAsync(user.Name);
-
+            if (user.Name != null) return await _userService.DeleteUserAsync(user.Name);
+            return new BadRequestObjectResult(new {massage = "Пользователь не обнаружен"});
         }
 
         [HttpPost("Register")]
         public async Task<IActionResult> Register([FromBody] RegisterDto model)
         {
-
-
             return await _userService.Register(model);
         }
 
         [HttpPost("Login")]
         public async Task<IActionResult> Login([FromBody] LoginDto model)
         {
-
-
             return await _userService.Login(model);
         }
 
         [HttpPut("RefreshToken")]
         public async Task<IActionResult> RefreshToken([FromBody] RefreshTokenDto model)
         {
-
-
             return await _userService.RefreshToken(model);
         }
 
@@ -84,8 +70,6 @@ namespace API.Controllers
         [HttpPost("RegisterOrLogin")]
         public async Task<IActionResult> RegisterOrLogin([FromBody] UserDto userDto)
         {
-
-
             return await _userService.RegisterOfLogin(Maps.MapUserDtoToRegistaerDto(userDto));
         }
     }
